Use as package (#35)
* Scruff can now be used as an include
* Removed dangling import in range_utils.jl
* Fixed bug in staticmodel.jl

---------
