+++ src/runtime.jl
@@ -20,7 +20,6 @@ export
     get_message,
     get_name,
     get_network,
-    get_range,
     get_sfunc,
     get_state,
     get_time,
@@ -31,7 +30,6 @@ export
     has_instance,
     has_intervention,
     has_previous_instance,
-    has_range,
     has_state,
     has_value,
     previous_instance,
@@ -51,7 +49,6 @@ export
     latest_instance_before,
     remove_messages!,
     set_message!,
-    set_range!,
     set_state!,
     set_time!,
     set_value!
@@ -60,7 +57,7 @@ using DataStructures
 using Dates
 using Random
 
-using Scruff.Models
+using .Models
 
 import Base.isless
 
@@ -559,72 +556,6 @@ function _unwrap(p)
     else
         p
     end
-end
-
-"""
-    RANGE
-
-The constant key used to store the range of a specific variable instance 
-"""
-const RANGE = :__range__
-
-"""
-    set_range!(runtime::Runtime, inst::Instance{O}, range::Vector{<:O}, depth::Int = 1) where O
-
-Sets the range value for the given instance. Defaults to depth of 1.
-"""
-function set_range!(runtime::Runtime, inst::Instance{O}, range::Vector{<:O}, depth::Int = 1) where O
-    if has_value(runtime, inst, RANGE)
-        curr = get_value(runtime, inst, RANGE)
-        s = Tuple{Int, Vector{O}}[]
-        i = 1
-        while i <= length(curr)
-            pair = curr[i]
-            d = pair[1]
-            if d > depth
-                push!(s, pair)
-                i += 1
-            end
-            push!(s, (depth, range))
-            i = d == depth ? i+1 : i
-            for j = i:length(curr)
-                push!(s, curr[j])
-            end
-            set_value!(runtime, inst, RANGE, s)
-            return
-        end
-        push!(s, (depth, range))
-        set_value!(runtime, inst, RANGE, s)
-    else
-        set_value!(runtime, inst, RANGE, [(depth, range)])
-    end
-end
-
-"""
-    get_range(runtime::Runtime, inst::Instance, depth = max_value(Int))
-
-Returns the range value for the given instance; this will return
-`nothing` if no range has been set.
-
-The depth specifies the maximum depth of range desired.
-"""
-function get_range(runtime::Runtime, inst::Instance, depth = typemax(Int))
-    has_range(runtime, inst, depth) || return nothing
-    rng = get_value(runtime, inst, RANGE)
-    for i in 1:length(rng)
-        (d,r) = rng[i]
-        if d <= depth
-            return r
-        end
-    end
-    return nothing
-end
-
-function has_range(runtime::Runtime, inst::Instance, depth::Int = typemax(Int)) 
-    has_value(runtime, inst, RANGE) || return false
-    r = get_value(runtime, inst, RANGE)
-    (d,_) = r[length(r)]
-    return d <= depth
 end
 
 """
